+++ gulpfile.js/tests/fixtures/components/component/example.js
@@ -1,1 +1,1 @@
-console.log('example')
+module.exports = 'JavaScript for example component'